% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions}
\alias{actions}
\title{Download Actions Data}
\usage{
actions(
  action_id = NULL,
  assessment_unit_id = NULL,
  state_code = NULL,
  organization_id = NULL,
  summarize = FALSE,
  parameter_name = NULL,
  pollutant_name = NULL,
  action_type_code = NULL,
  agency_code = NULL,
  pollutant_source_code = NULL,
  action_status_code = NULL,
  completion_date_later_than = NULL,
  completion_date_earlier_than = NULL,
  tmdl_date_later_than = NULL,
  tmdl_date_earlier_then = NULL,
  last_change_later_than_date = NULL,
  last_change_earlier_than_date = NULL,
  return_count_only = FALSE,
  tidy = TRUE,
  ...
)
}
\arguments{
\item{action_id}{(character) specifies what action to retrieve. multiple values allowed. optional}

\item{assessment_unit_id}{(character)}

\item{state_code}{(character)}

\item{organization_id}{(character)}

\item{summarize}{(logical)}

\item{parameter_name}{(character)}

\item{pollutant_name}{(character)}

\item{action_type_code}{(character)}

\item{agency_code}{(character)}

\item{pollutant_source_code}{(character)}

\item{action_status_code}{(character)}

\item{completion_date_later_than}{(character)}

\item{completion_date_earlier_than}{(character)}

\item{tmdl_date_later_than}{(character)}

\item{tmdl_date_earlier_then}{(character)}

\item{last_change_later_than_date}{(character)}

\item{last_change_earlier_than_date}{(character)}

\item{return_count_only}{(logical)}

\item{tidy}{(logical) \code{TRUE} (default) the function returns a tidied tibble. \code{FALSE} the function returns the raw JSON string.}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
tibble
}
\description{
Provides data about actions (TMDLs, 4B Actions, Alternative Actions, Protection Approach Actions) that have been finalized.
}
