% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{actions_to_tibble}
\alias{actions_to_tibble}
\title{Convert Action JSON to Tibble}
\usage{
actions_to_tibble(content, count = FALSE, summarize = FALSE)
}
\arguments{
\item{content}{json}

\item{count}{logical}

\item{summarize}{character}
}
\description{
Convert Action JSON to Tibble
}
\keyword{internal}
