% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment_units.R
\name{assessment_units}
\alias{assessment_units}
\title{Download Assessment Unit Summary}
\usage{
assessment_units(
  assessment_unit_identifer = NULL,
  state_code = NULL,
  organization_id = NULL,
  epa_region = NULL,
  huc = NULL,
  county = NULL,
  assessment_unit_name = NULL,
  last_change_later_than_date = NULL,
  last_change_earlier_than_date = NULL,
  status_indicator = NULL,
  return_count_only = NULL,
  ...
)
}
\arguments{
\item{assessment_unit_identifer}{(character) Filters the list of assessment units to one or more specific assessment units. Multiple values can be provided.}

\item{state_code}{(character) Filters the list of assessment units to only those having a state code matches one in the provided list of states. Multiple values can be provided.}

\item{organization_id}{(character) Filters the list of assessment units to only those having an organization ID that matches one in the provided list of IDs. Multiple values can be provided.}

\item{epa_region}{(character) Filters the list of assessment units to only those having an EPA region that matches one in the provided list of regions. Multiple values can be provided.}

\item{huc}{(character) Filters the list of assessment units to only those which have a location type of HUC and the location value matches one in the provided list of HUCs. Multiple values can be provided.}

\item{county}{(character) Filters the list of assessment units to only those which have a location type of county and the location value matches one in the provided list of counties. Multiple values can be provided.}

\item{assessment_unit_name}{(character) Filters the list of assessment units to only those having an assessment unit name matching the provided value.}

\item{last_change_later_than_date}{(character) yyyy-mm-dd}

\item{last_change_earlier_than_date}{(character) yyyy-mm-dd}

\item{status_indicator}{(character) "A" for active, "R" for retired. optional}

\item{return_count_only}{(character) "Y" for yes, "N" for no. Defaults to "N". optional}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{

}
\description{
Download Assessment Unit Summary
}
