% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huc12_summary.R
\name{huc12_summary}
\alias{huc12_summary}
\title{Download HUC12 Summary}
\usage{
huc12_summary(huc, tidy = TRUE, ...)
}
\arguments{
\item{huc}{(character) 12-digit hydrologic unit code. required}

\item{tidy}{(logical) \code{TRUE} (default) the function returns a tidied tibble. \code{FALSE} the function returns the raw JSON string.}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
a list of tibbles
}
\description{
Provides summary data for a 12-digit Hydrologic Unit Code (HUC12), based on
Assessment Units in the HUC12. Watershed boundaries may cross state
boundaries, so the service may return assessment units from multiple
organizations. Returns the assessment units in the HUC12, size and
percentages of assessment units considered Good, Unknown, or Impaired.
}
