% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rATTAINS-package.R
\docType{package}
\name{rATTAINS-package}
\alias{rATTAINS}
\alias{rATTAINS-package}
\title{rATTAINS: Access EPA 'ATTAINS' Data}
\description{
An R interface to United States Environmental Protection Agency (EPA) Assessment, Total Maximum Daily Load (TMDL) Tracking and Implementation System ('ATTAINS') data. 'ATTAINS' is the EPA database used to track information provided by states about water quality assessments conducted under federal Clean Water Act requirements. ATTAINS information and API information is available at \url{https://www.epa.gov/waterdata/attains}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mps9506/rATTAINS}
  \item \url{https://mps9506.github.io/rATTAINS/}
  \item Report bugs at \url{https://github.com/mps9506/rATTAINS/issues}
}

}
\author{
\strong{Maintainer}: Michael Schramm \email{mpschramm@gmail.com} (\href{https://orcid.org/0000-0003-1876-6592}{ORCID}) [copyright holder]

}
\keyword{internal}
