% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rATTAINS_opts.R
\name{rATTAINS_options}
\alias{rATTAINS_options}
\title{rATTAINS options}
\usage{
rATTAINS_options(cache_downloads = FALSE)
}
\arguments{
\item{cache_downloads}{(logical) whether to locally cache downloads. default: \code{FALSE}}
}
\value{
\value{None}
}
\description{
rATTAINS options
}
\details{
rATTAINS package level options; stored in an internal
package environment \code{rATTAINSenv}
}
\examples{
\dontrun{
rATTAINS_options(cache_downloads = FALSE)
}
}
\seealso{
\link{rATTAINS_caching} for managing cached files
}
