% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessments.R
\name{assessments}
\alias{assessments}
\title{Download Assessment Decisions}
\usage{
assessments(
  assessment_unit_id = NULL,
  state_code = NULL,
  organization_id = NULL,
  reporting_cycle = NULL,
  use = NULL,
  use_support = NULL,
  parameter = NULL,
  parameter_status_name = NULL,
  probable_source = NULL,
  agency_code = NULL,
  ir_category = NULL,
  state_ir_category_code = NULL,
  multicategory_search = NULL,
  last_change_later_than_date = NULL,
  last_change_earlier_than_date = NULL,
  return_count_only = FALSE,
  exclude_assessments = FALSE,
  ...
)
}
\arguments{
\item{assessment_unit_id}{(character)}

\item{state_code}{(character)}

\item{organization_id}{(character)}

\item{reporting_cycle}{(character)}

\item{use}{(character)}

\item{use_support}{(character)}

\item{parameter}{(character)}

\item{parameter_status_name}{(character)}

\item{probable_source}{(character)}

\item{agency_code}{(character)}

\item{ir_category}{(character)}

\item{state_ir_category_code}{(character)}

\item{multicategory_search}{(character)}

\item{last_change_later_than_date}{(character)}

\item{last_change_earlier_than_date}{(character)}

\item{return_count_only}{(logical)}

\item{exclude_assessments}{(logical)}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{

}
\description{
Download Assessment Decisions
}
