% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_values.R
\name{domain_values}
\alias{domain_values}
\title{Download Domain Values}
\usage{
domain_values(domain_name = NULL, context = NULL, ...)
}
\arguments{
\item{domain_name}{(character)}

\item{context}{(character)}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{

}
\description{
Download Domain Values
}
