% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_values.R
\name{domain_values}
\alias{domain_values}
\title{Download Domain Values}
\usage{
domain_values(domain_name = NULL, context = NULL, tidy = TRUE, ...)
}
\arguments{
\item{domain_name}{(character)}

\item{context}{(character)}

\item{tidy}{(logical) \code{TRUE} (default) the function returns a tidied tibble. \code{FALSE} the function returns the raw JSON string.}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
tibble or character string
}
\description{
Download Domain Values
}
