% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_summary.R
\name{state_summary}
\alias{state_summary}
\title{Download State Summaries}
\usage{
state_summary(organization_id, reporting_cycle = NULL, ...)
}
\arguments{
\item{organization_id}{(character) Organization identifier used by EPA.
required}

\item{reporting_cycle}{(character) 4 digit reporting cycle year. Typically
even numbered years. Will return reporting data for all years prior to and
including the reporting cycle by reporting cycle. optional}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
tibble
}
\description{
Provides summary information for assess uses for an organization (State,
Territory or Tribe) and Integrated Reporting Cycle. The Organization ID for
the state, territory or tribe is required. If a Reporting Cycle isn't
provided, the service will return the most recent cycle. If a reporting Cycle
is provided, the service will return a summary for the requested cycle.
}
