% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{surveys}
\alias{surveys}
\title{Download State Survey Results}
\usage{
surveys(organization_id = NULL, survey_year = NULL, tidy = TRUE, ...)
}
\arguments{
\item{organization_id}{(character) - Filters the list to only those “belonging to” one of the specified organizations. Multiple values may be specified. required}

\item{survey_year}{(character) Filters the list to the year the survey was performed. optional.}

\item{tidy}{(logical) \code{TRUE} (default) the function returns a tidied tibble. \code{FALSE} the function returns the raw JSON string.}

\item{...}{list of curl options passed to \code{\link[crul:HttpClient]{crul::HttpClient()}}}
}
\value{
a list of tibbles
}
\description{
Downloads data about state statistical (probability) survey results.
}
\note{
Arguments that allow multiple values should be entered as a comma separated string with no spaces (\code{organization_id = "DOEE,21AWIC"}).
}
